+++ src/index.ts
@@ -12,20 +12,19 @@ export * from './jsii-project';
 export * from './json';
 export * from './license';
 export * from './makefile';
-export * from './next';
 export * from './node-project';
 export * from './project';
-export * from './react';
+export * from './sample-file';
 export * from './semver';
 export * from './start';
+export * from './textfile';
 export * from './toml';
 export * from './typescript';
 export * from './version';
-export * from './textfile';
-export * from './sample-file';
 export * from './yaml';
 
 // export submobules
 export * as seqs from './seqs';
+export * as web from './web';
 export * as github from './github';
 export * as vscode from './vscode';