+++ CHANGELOG.md
@@ -2,6 +2,27 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### 0.5.4 (2020-11-28)
+
+
+### Features
+
+* create github repository ([#336](https://github.com/projen/projen/issues/336)) ([5590e08](https://github.com/projen/projen/commit/5590e0838b202de1902f1ab697be5296192bb85d)), closes [#310](https://github.com/projen/projen/issues/310)
+
+### 0.5.3 (2020-11-26)
+
+
+### Features
+
+* **nextjs:** add tailwindcss option ([#326](https://github.com/projen/projen/issues/326)) ([e02eb65](https://github.com/projen/projen/commit/e02eb650256b2a3f6421b42629ed8e5e4d654642)), closes [#325](https://github.com/projen/projen/issues/325)
+
+### 0.5.2 (2020-11-25)
+
+
+### Bug Fixes
+
+* **new:** cannot pass string paramaeters as CLI switches ([#335](https://github.com/projen/projen/issues/335)) ([db51548](https://github.com/projen/projen/commit/db515481e14a59ec014209e082c57496a56dc769)), closes [#305](https://github.com/projen/projen/issues/305)
+
 ### 0.5.1 (2020-11-24)